Remove warnings about missing imports and move integration tests into proper place
